requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
authx==1.4.1
bcrypt==4.3.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
cryptography==44.0.2
distlib==0.3.9
dulwich==0.22.8
ecdsa==0.19.1
fastapi==0.115.11
fastapi-csrf-protect==1.0.2
fastapi-jwt-auth==0.5.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
keyring==25.6.0
Mako==1.3.9
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
pbs-installer==2025.3.17
pkginfo==1.12.1.2
platformdirs==4.3.7
pluggy==1.5.0
poetry==2.1.1
poetry-core==2.1.1
propcache==0.3.0
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==1.7.1
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.2
pywin32-ctypes==0.2.3
RapidFuzz==3.12.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
text-unidecode==1.3
tomlkit==0.13.2
trove-classifiers==2025.3.19.19
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
yarl==1.18.3
zstandard==0.23.0
